refactor : 전역 예외처리 관련 파일 application 계층으로 이동 (#23)
