+++ order/src/main/java/com/sparta/hotdeal/order/application/service/order/OrderBasketService.java
@@ -3,8 +3,8 @@ package com.sparta.hotdeal.order.application.service.order;
 import com.sparta.hotdeal.order.domain.entity.AuditingDate;
 import com.sparta.hotdeal.order.domain.entity.basket.Basket;
 import com.sparta.hotdeal.order.domain.repository.BasketRepository;
-import com.sparta.hotdeal.order.infrastructure.exception.ApplicationException;
-import com.sparta.hotdeal.order.infrastructure.exception.ErrorCode;
+import com.sparta.hotdeal.order.application.exception.ApplicationException;
+import com.sparta.hotdeal.order.application.exception.ErrorCode;
 import java.util.List;
 import java.util.UUID;
 import lombok.RequiredArgsConstructor;
