+++ order/src/main/java/com/sparta/hotdeal/order/application/service/order/OrderProductService.java
@@ -7,8 +7,8 @@ import com.sparta.hotdeal.order.domain.entity.basket.Basket;
 import com.sparta.hotdeal.order.domain.entity.order.Order;
 import com.sparta.hotdeal.order.domain.entity.order.OrderProduct;
 import com.sparta.hotdeal.order.domain.repository.OrderProductRepository;
-import com.sparta.hotdeal.order.infrastructure.exception.ApplicationException;
-import com.sparta.hotdeal.order.infrastructure.exception.ErrorCode;
+import com.sparta.hotdeal.order.application.exception.ApplicationException;
+import com.sparta.hotdeal.order.application.exception.ErrorCode;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
